# openapi-gestion.yaml
...
openapi: 3.0.3
info:
  title: PID Gijón — API de gestión
  version: 1.0.0
  description: 'API de lectura y gestión de la Plataforma de Integración de Datos del Ayuntamiento de
    Gijón.


    El dato **no entra por aquí**: la ingesta va por el núcleo FIWARE (sensor → IoT Agent → Orion-LD →
    RabbitMQ → consumidores). Esta API cubre entidades, dispositivos, series temporales, cuadros de mando,
    alarmas y preferencias.


    Autenticación por *bearer token* JWT emitido por Keycloak. La API no participa en la sesión: el
    token se obtiene y se renueva contra el proveedor de identidad (código de autorización + PKCE para
    clientes de navegador). Las rutas bajo `/internal` no llevan token: solo son accesibles desde la
    red interna y las consume el gateway APISIX para validar permiso de escritura por tenant.'
  license:
    name: EUPL-1.2
    url: https://joinup.ec.europa.eu/collection/eupl/eupl-text-eupl-12
servers:
- url: /api
  description: Ruta base del servicio
security:
- bearerAuth: []
...
paths:
...
  /V1/panels:
    get:
      tags:
      - Cuadros de mando
      summary: Index
      operationId: Panel_index_get
      responses:
        '200':
          description: OK
        '401':
          description: No autenticado
        '403':
          description: Sin permiso
    post:
      tags:
      - Cuadros de mando
      summary: Store
      operationId: Panel_store_post
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
      responses:
        '201':
          description: Creado
        '401':
          description: No autenticado
        '403':
          description: Sin permiso
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT emitido por Keycloak (flujo de código con PKCE). Se obtiene y se renueva contra el proveedor de identidad.